pyproject.toml
[tool.poetry]
name = "fastapi-crud-app"
version = "0.1.0"
description = ""
authors = ["Vadym <vrupa1713@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.109.0"
uvicorn = {extras = ["standard"], version = "^0.27.0.post1"}
pydantic = {extras = ["email"], version = "^2.6.0"}
python-multipart = "^0.0.6"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
SQLAlchemy = "^2.0.25"
alembic = "^1.13.1"
shortuuid = "^1.0.11"
python-dotenv = "^1.0.1"
pydantic-settings = "^2.1.0"
mysqlclient = "^2.2.1"
pyjwt = "^2.8.0"
python-decouple = "^3.8"
bcrypt = "4.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
